requirements.txt
flet>=0.21.0
pyserial>=3.5
numpy>=1.21.0
pandas>=2.0.0
openpyxl>=3.1.0